Refactor IR
