+++ src/ir.rs
@@ -1,7 +1,5 @@
 mod build;
 mod configuration;
-mod rule;
 
 pub use build::*;
 pub use configuration::*;
-pub use rule::*;
